+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
+- CO-2461 - add clear cache button, add disable cache button
+- fix debug log setting
 
 ## 1.40.4 _2023-11-06_
 - CO-2458 - consider attribute IDs during product pull