+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
 - CO-2471 - update argument order in image push
+- CO-2461 - add clear cache button, add disable cache button
+- fix debug log setting
+
+## 1.40.4 _2023-11-06_
 - CO-2458 - consider attribute IDs during product pull
 - Fix Customer Group Prices